chore - Proguard clean - Moved DTOs to common folder - Add proper proguard rules
