+++ app/proguard-rules.pro
@@ -29,3 +29,8 @@
 -dontwarn com.squareup.javawriter.JavaWriter
 # Uncomment this if you use Mockito
 -dontwarn org.mockito.**
+
+-dontnote okhttp3.**, okio.**, retrofit2.**
+-dontwarn retrofit2.**
+-keep class retrofit2.** { *; }
+-keep class com.aggarwalankur.stationdata.network.dto** {*;}
